# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineLoop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Film {
  id: number;
  title: string;
  description: string;
  genre: string;
  thumbnailUrl: string;
  creatorName: string;
  creatorAvatar: string;
  totalEpisodes: number;
  episodesReleased: number;
  totalLikes: number;
  totalVotes: number;
  isSerialised: boolean;
  createdAt: string;
}

export interface CreateFilmInput {
  title: string;
  description: string;
  genre: string;
  thumbnailUrl: string;
  isSerialised: boolean;
}

export interface Episode {
  id: number;
  filmId: number;
  episodeNumber: number;
  title: string;
  description: string;
  videoUrl: string;
  thumbnailUrl: string;
  durationSeconds: number;
  isLocked: boolean;
  unlockCondition?: string;
  likes: number;
  cliffhangerText?: string;
  hasActivePoll: boolean;
  createdAt: string;
}

export interface LikeResponse {
  liked: boolean;
  totalLikes: number;
  xpEarned: number;
}

export interface VoteOption {
  id: number;
  text: string;
  voteCount: number;
  percentage: number;
}

export interface VotePoll {
  id: number;
  episodeId: number;
  filmTitle: string;
  question: string;
  options: VoteOption[];
  totalVotes: number;
  isActive: boolean;
  endsAt?: string;
  userVoted: boolean;
  userChoice?: number;
}

export interface CastVoteInput {
  optionId: number;
}

export interface VoteResult {
  poll: VotePoll;
  xpEarned: number;
  newTotal: number;
}

export interface User {
  id: number;
  username: string;
  avatarUrl: string;
  level: number;
  xp: number;
  xpToNextLevel: number;
  streakDays: number;
  totalVotes: number;
  totalWatched: number;
  totalLikes: number;
  badges: string[];
  followedCharacters: number[];
}

export interface XpEvent {
  id: number;
  action: string;
  xpAmount: number;
  description: string;
  createdAt: string;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  lastActiveDate: string;
  streakBonusXp: number;
  isStreakAlive: boolean;
}

export interface Character {
  id: number;
  name: string;
  filmTitle: string;
  avatarUrl: string;
  bio: string;
  followerCount: number;
  isFollowed: boolean;
  latestUpdate: string;
}

export interface FollowResponse {
  following: boolean;
  followerCount: number;
}

export interface FeedItem {
  id: number;
  type: string;
  episode: Episode;
  film: Film;
  activePoll?: VotePoll;
}

export type TrendingFeedRisingCreatorsItem = {
  name: string;
  avatarUrl: string;
  followerGrowth: number;
};

export interface TrendingFeed {
  topFilms: Film[];
  hotPolls: VotePoll[];
  risingCreators: TrendingFeedRisingCreatorsItem[];
}

export type PlatformStatsTopGenresItem = {
  genre: string;
  count: number;
};

export interface PlatformStats {
  totalFilms: number;
  totalEpisodes: number;
  activeVotePolls: number;
  totalVotesCast: number;
  totalUsers: number;
  topGenres: PlatformStatsTopGenresItem[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  avatarUrl: string;
  level: number;
  xp: number;
  streakDays: number;
}

export type GetFilmsParams = {
  genre?: string;
  limit?: number;
  offset?: number;
};

export type GetFeedParams = {
  limit?: number;
  offset?: number;
};