# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineLoop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: number;
  name: string;
  filmTitle: string;
  avatarUrl: string;
  bio: string;
  followerCount: number;
  isFollowed: boolean;
  latestUpdate: string;
}