# lib/api-zod/src/generated/types/episode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineLoop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Episode {
  id: number;
  filmId: number;
  episodeNumber: number;
  title: string;
  description: string;
  videoUrl: string;
  thumbnailUrl: string;
  durationSeconds: number;
  isLocked: boolean;
  unlockCondition?: string;
  likes: number;
  cliffhangerText?: string;
  hasActivePoll: boolean;
  createdAt: string;
}