# lib/api-zod/src/generated/types/film.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineLoop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Film {
  id: number;
  title: string;
  description: string;
  genre: string;
  thumbnailUrl: string;
  creatorName: string;
  creatorAvatar: string;
  totalEpisodes: number;
  episodesReleased: number;
  totalLikes: number;
  totalVotes: number;
  isSerialised: boolean;
  createdAt: string;
}