# lib/api-zod/src/generated/types/streakInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineLoop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  lastActiveDate: string;
  streakBonusXp: number;
  isStreakAlive: boolean;
}