# lib/api-zod/src/generated/types/votePoll.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineLoop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoteOption } from "./voteOption";

export interface VotePoll {
  id: number;
  episodeId: number;
  filmTitle: string;
  question: string;
  options: VoteOption[];
  totalVotes: number;
  isActive: boolean;
  endsAt?: string;
  userVoted: boolean;
  userChoice?: number;
}