# lib/api-zod/src/generated/types/xpEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineLoop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface XpEvent {
  id: number;
  action: string;
  xpAmount: number;
  description: string;
  createdAt: string;
}